Tighter mapping of input values to output values.
I dumped out some raw radio inputs and PPM outputs to try to get things
a bit closer to expected range.

Consider this example input:

    | Ch | Input | Output |
    |----+-------+--------|
    |  1 |  1482 |    992 |
    |  2 |  2257 |   1512 |
    |  3 |  2244 |   1503 |
    |  4 |  2256 |   1511 |
    |  5 |  1502 |   1006 |
    |  6 |  2250 |   1507 |
    |  7 |  3018 |   2022 |
    |  8 |  2246 |   1504 |

Channel 1 was throttle and at its lowest position.  Channel 7 was S2
turned all the way up to 100%.  Most channels were hovering around their
middle positions.

The preexisting mapping is basically `input * 0.67`.  This is reasonably
close, but is a bit off in the middle, too low on the lower end and too
high on the upper (ranges from 992μS to 2022μS with the middle around
1507μS).

With the observed low end raw data from the radio at 1482 (0x5ca) and
the upper end at 3018 (0xbca), I attempted to map these directly to the
desired 1000μS-2000μS using the map() function:

    map(ci, 1482, 3018, 1000, 2000);

Curious about the complexity of this, I dumped the assembly output and
found it to be a... lot for a fairly frequent and critical loop, so I
simplified it:

    (x - in_min) * (out_max - out_min + 1) / (in_max - in_min + 1) + out_min;
    (x - 1482) * (2000 - 1000 + 1) / (3018 - 1482 + 1) + 1000;
    (x - 1482) * (1001) / (1537) + 1000;
    (x - 1482) * .6512687 + 1000;

This yielded significantly fewer instructions, but overall pretty good
results:

    | Ch | Input | Output |
    |----+-------+--------|
    |  1 |  1482 |   1000 |
    |  2 |  2257 |   1505 |
    |  3 |  2244 |   1496 |
    |  4 |  2256 |   1504 |
    |  5 |  1502 |   1013 |
    |  6 |  2250 |   1500 |
    |  7 |  3018 |   2000 |
    |  8 |  2246 |   1497 |

With one addition and subtraction more than the previous implementation,
we're getting better results, staying closer to the expected values.